Introduce ApiReplacementMessageBuilder
All resource strings for MemberAccess analyzers can now
be obtained by methods from this static class. Strings are parametrized.
This ensures common same messages for all similair errors so that they
can be changed easily and are conformant with one another.
Static class for Messages for COdeFixes was also added.