Introduce BaseMemberInvocationAnalyzer and fix WhereLike
This is needed because BaseMemberAccessAnalyzer class does not reflect the
invocation of method. Therefore in something like
whereCondition.Or().WhereLike(...).Or(), it was very hard to locate member
access which is actually forbidden.
With new Base class for invocatoins only, intellisense in runtime is
better as it put squiggles more precisely (only under the method call).
Moreover it helpes when codefix has to work with method arguments like in
ClientScriptManager.