Perf - fix version number of AnalyzersVersions nuget
This is to ease testing so that there is no need to upgrade the installed
version of nuget when a change in source is needed. Note that this project
is only for testing purposes and won't be part of final package, so it
should not be an issue to have fixed version number.